# specs/048-tenant-isolation/contracts/csp-inherited-components.openapi.yaml
openapi: 3.0.3
info:
  title: ATO Copilot — CSP-Inherited Components API (US9)
  version: 1.0.0
  description: |
    Endpoints that ingest a CSP's existing ATO artifacts (PDF SSP, DOCX, OSCAL JSON,
    FedRAMP / eMASS XLSX, eMASS ZIP), extract candidate components, attempt to
    auto-map them to NIST 800-53 capabilities, and expose the resulting
    CSP-inherited components read-only to every hosted tenant.

    Two surfaces are defined:
      * `/api/csp/onboarding/atos/*` — wizard step (FR-099..FR-103, MultiTenant only)
      * `/api/csp/inherited-components/*` — post-onboarding management (FR-104..FR-106)

    All write endpoints require role `CSP.Admin`. All endpoints require a valid
    Entra-issued bearer token. In `SingleTenant` mode the entire surface returns
    `404` (`SINGLE_TENANT_MODE`).
servers:
  - url: /api
    description: Default host
...
paths:
  /csp/onboarding/atos/upload:
    post:
      tags: [csp-onboarding-atos]
      summary: Upload one or more ATO documents during the CSP wizard
      description: |
        Multipart upload, one or more files. Each file is parsed using the parser
        appropriate for its content type, candidate components are extracted as
        `CspInheritedComponent` rows (`Status = Draft`), and the AI capability
        mapping pipeline is invoked per component to produce `CspInheritedCapability`
        rows.

        * Max file size: 50 MB per file (`413 ATO_DOCUMENT_TOO_LARGE` if exceeded).
        * Supported content types: `application/pdf`, `application/vnd.openxmlformats-officedocument.wordprocessingml.document`, `application/json` (OSCAL SSP), `application/vnd.openxmlformats-officedocument.spreadsheetml.sheet`, `application/zip` (eMASS package).
        * Unknown content types: `400 UNSUPPORTED_ATO_DOCUMENT`.
        * Parser failures: `400 PARSE_FAILED` with the failing parser name.
        * If the AI mapping service is unreachable, components are still persisted but `aiMappingAvailable = false` in the response and **zero** capabilities are auto-created (operator can re-run via `/csp/inherited-components/{componentId}/remap`).
      operationId: cspOnboardingUploadAtos
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [files]
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
                  description: One or more ATO documents (max 50 MB each).
      responses:
        '200':
          description: Upload + parsing + (best-effort) AI mapping completed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AtoUploadResponse'
        '400':
          description: Unsupported content type or parse failure.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Caller is not `CSP.Admin`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Deployment is in `SingleTenant` mode (`SINGLE_TENANT_MODE`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '413':
          description: A file exceeds the 50 MB limit.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /csp/onboarding/atos/state:
    get:
      tags: [csp-onboarding-atos]
      summary: Retrieve the current ATO Documents step state for the wizard
      description: |
        Returns the running tally surfaced by FR-103 — `documentsUploaded`,
        `componentsExtracted`, `capabilitiesMapped`, `capabilitiesNeedsReview`,
        and `aiMappingAvailable` — plus a per-document list. Re-entrant.
      operationId: cspOnboardingAtoState
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current wizard state for the ATO Documents step.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AtoStepState'
        '403': { description: Not CSP-Admin., content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
        '404': { description: SingleTenant mode., content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SourceFormat:
      type: string
      enum: [Pdf, Docx, OscalJson, Xlsx, EmassZip, Manual]
...
    AtoUploadResponse:
      type: object
      required: [documentsAccepted, componentsExtracted, capabilitiesMapped, capabilitiesNeedsReview, aiMappingAvailable, files]
      properties:
        documentsAccepted: { type: integer, minimum: 0 }
        componentsExtracted: { type: integer, minimum: 0 }
        capabilitiesMapped: { type: integer, minimum: 0 }
        capabilitiesNeedsReview: { type: integer, minimum: 0 }
        aiMappingAvailable: { type: boolean }
        aiMappingFailureReason: { type: string, nullable: true }
        files:
          type: array
          items:
            type: object
            required: [fileName, sourceFormat, parsedSuccessfully, componentsExtracted]
            properties:
              fileName: { type: string }
              sourceFormat: { $ref: '#/components/schemas/SourceFormat' }
              parsedSuccessfully: { type: boolean }
              parseError: { type: string, nullable: true }
              componentsExtracted: { type: integer, minimum: 0 }
              capabilitiesMapped: { type: integer, minimum: 0 }
              capabilitiesNeedsReview: { type: integer, minimum: 0 }
    AtoStepState:
      type: object
      required: [documentsUploaded, componentsExtracted, capabilitiesMapped, capabilitiesNeedsReview, aiMappingAvailable]
      properties:
        documentsUploaded: { type: integer, minimum: 0 }
        componentsExtracted: { type: integer, minimum: 0 }
        capabilitiesMapped: { type: integer, minimum: 0 }
        capabilitiesNeedsReview: { type: integer, minimum: 0 }
        aiMappingAvailable: { type: boolean }
        aiMappingFailureReason: { type: string, nullable: true }
        documents:
          type: array
          items:
            type: object
            required: [fileName, sourceFormat, componentsExtracted]
            properties:
              fileName: { type: string }
              sourceFormat: { $ref: '#/components/schemas/SourceFormat' }
              componentsExtracted: { type: integer, minimum: 0 }
              capabilitiesMapped: { type: integer, minimum: 0 }
              capabilitiesNeedsReview: { type: integer, minimum: 0 }
...
    ErrorEnvelope:
      type: object
      required: [status, errorCode]
      properties:
        status: { type: integer }
        errorCode:
          type: string
          enum:
            - UNSUPPORTED_ATO_DOCUMENT
            - PARSE_FAILED
            - ATO_DOCUMENT_TOO_LARGE
            - CSP_ONBOARDING_INCOMPLETE
            - FORBIDDEN_NOT_CSP_ADMIN
            - SINGLE_TENANT_MODE
            - VALIDATION_FAILED
            - AI_MAPPING_UNAVAILABLE
        message: { type: string }
        suggestion: { type: string }
        details:
          type: object
          additionalProperties: true